# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZTAP REST API
  version: v1
  description: |
    Minimal REST API for ZTAP.

    Auth uses bearer tokens from `POST /v1/auth/login`.
servers:
  - url: http://127.0.0.1:8080
...
paths:
...
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Authenticate and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/auth/whoami:
    get:
      tags: [auth]
      summary: Return the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhoAmIResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/status:
    get:
      tags: [status]
      summary: Process status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/enforcement/status:
    get:
      tags: [enforcement]
      summary: Enforcement status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Enforcement status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnforcementStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/enforcement/start:
    post:
      tags: [enforcement]
      summary: Start enforcement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnforcementStartRequest"
      responses:
        "200":
          description: Enforcement started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnforcementStartResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/enforcement/stop:
    post:
      tags: [enforcement]
      summary: Stop enforcement
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                type: object
                properties:
                  stopped:
                    type: boolean
                required: [stopped]
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/flows/stream:
    get:
      tags: [flows]
      summary: Stream flow events (Server-Sent Events)
      description: |
        Returns a `text/event-stream` where each `data:` line contains a JSON-encoded flow event.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/config/backup:
    post:
      tags: [config]
      summary: Download a backup bundle
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupRequest"
      responses:
        "200":
          description: Backup bundle (.tar.gz)
          content:
            application/gzip:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/config/restore:
    post:
      tags: [config]
      summary: Restore from a backup bundle
      description: |
        Upload a `.tar.gz` created by `/v1/config/backup`.
      security:
        - bearerAuth: []
      parameters:
        - name: dry_run
          in: query
          schema:
            type: boolean
          description: If true, compute restore plan without applying.
        - name: force
          in: query
          schema:
            type: boolean
          description: If true, allow destructive overwrites.
      requestBody:
        required: true
        content:
          application/gzip:
            schema:
              type: string
              format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Restore plan/report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Force required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Bundle too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/compliance/report:
    post:
      tags: [compliance]
      summary: Generate a compliance report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceReportRequest"
      responses:
        "200":
          description: Compliance report
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/compliance/export:
    post:
      tags: [compliance]
      summary: Export compliance report (json/csv)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceExportRequest"
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/policies:
    get:
      tags: [policies]
      summary: List policies
      security:
        - bearerAuth: []
      parameters:
        - name: tenant
          in: query
          schema:
            type: string
          description: Optional tenant filter
      responses:
        "200":
          description: Policy list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoliciesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /v1/users:
    get:
      tags: [users]
      summary: List users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    post:
      tags: [users]
      summary: Create a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /v1/cluster/status:
    get:
      tags: [cluster]
      summary: Cluster status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cluster status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          description: Cluster not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/cluster/nodes:
    get:
      tags: [cluster]
      summary: List cluster nodes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Nodes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterNodesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          description: Cluster not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    post:
      tags: [cluster]
      summary: Register a cluster node
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterNodeRequest"
      responses:
        "201":
          description: Node registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterNode"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          description: Cluster not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthLoginResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
        role:
          type: string
        expires_at:
          type: string
          format: date-time
      required: [token, username, role, expires_at]
    WhoAmIResponse:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
        expires_at:
          type: string
          format: date-time
      required: [username, role, expires_at]
    StatusResponse:
      type: object
      properties:
        os:
          type: string
        arch:
          type: string
        hostname:
          type: string
        pid:
          type: integer
          format: int32
        uptime:
          type: string
      required: [os, arch, hostname, pid, uptime]
    EnforcementStatusResponse:
      type: object
      properties:
        platform:
          type: string
        ebpf_active:
          type: boolean
        flow_events_pin_path:
          type: string
      required: [platform, ebpf_active, flow_events_pin_path]
...
    EnforcementStartResponse:
      type: object
      properties:
        enforced:
          type: boolean
        platform:
          type: string
      required: [enforced, platform]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    MethodNotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"